Clean up axios get request and useEffect in RatingsAndReviews and ReviewsList
